relaciones modelos, se agregaron tablas y modelos morfos Logs, Settings y Files

+++ database/migrations/tenant/2023_07_03_210419_create_quotation_details_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('quotation_details', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('quotation_id');
-            $table->integer('product_or_service_id');
+            $table->unsignedBigInteger('quotation_id');
+            $table->unsignedBigInteger('product_or_service_id');
             $table->date('date');
             $table->text('description', 500)->nullalble();
             $table->decimal('quantity', 30, 6);
