+++ database/migrations/tenant/2023_07_03_213025_create_invoices_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('invoices', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('customer_id');
+            $table->unsignedBigInteger('customer_id');
             $table->date('date');
             $table->decimal('amount', 30, 6);
             $table->decimal('discount', 30, 6);
