+++ database/migrations/tenant/2023_07_06_113559_create_customer_credit_notes_documents_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('customer_credit_notes_documents', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('customer_credit_note_id');
-            $table->integer('invoice_id');
+            $table->unsignedBigInteger('customer_credit_note_id');
+            $table->unsignedBigInteger('invoice_id');
             $table->string('UUID', 255);
             $table->decimal('total', 30, 6);
             $table->decimal('discount', 30, 6);
